recommender/pyproject.toml
[project]
name = "recommender"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)"
]

[tool.poetry]
packages = [{include = "recommender", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
